requirements.txt
apache-beam==2.48.0
apache-flink==1.18.0
apache-flink-libraries==1.18.0
avro-python3==1.10.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.4.2
docopt==0.6.2
fastavro==1.9.2
fasteners==0.19
finnhub-python==2.4.19
gevent==23.9.1
greenlet==3.0.3
grpcio==1.60.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.6
kafka-python==2.0.2
numpy==1.24.4
oauthlib==3.2.2
objsize==0.6.1
orjson==3.9.10
pandas==2.1.4
parquet-cli==1.3
proto-plus==1.23.0
protobuf==4.23.4
py4j==0.10.9.7
pyarrow==11.0.0
pycparser==2.21
pydot==1.4.2
pymongo==4.6.1
pyparsing==3.1.1
pyspark==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
tweepy==4.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
websocket==0.2.1
websocket-client==1.7.0
zope.event==5.0
zope.interface==6.1
zstandard==0.22.0
